Go ES2017; distribute bundles in multiple formats
Builds UMD and ES modules for ES2017 targets, and a UMD bundle for ES5 targets.

      -rw-r--r--  1 sam  staff   83655 Dec 15 15:40 dist/turbo.es2017-esm.js
      -rw-r--r--  1 sam  staff   94086 Dec 15 15:40 dist/turbo.es2017-umd.js
      -rw-r--r--  1 sam  staff  129575 Dec 15 15:40 dist/turbo.es5-umd.js

        "module": "dist/turbo.es2017-esm.js",
        "main": "dist/turbo.es5-umd.js",
