+++ package.json
@@ -2,9 +2,9 @@
   "name": "@hotwired/turbo",
   "version": "7.0.0-beta.1",
   "description": "Turbo makes navigating your web application faster",
-  "module": "dist/index.js",
-  "main": "dist/turbo.js",
-  "types": "dist/index.d.ts",
+  "module": "dist/turbo.es2017-esm.js",
+  "main": "dist/turbo.es5-umd.js",
+  "types": "dist/types/index.d.ts",
   "files": [
     "dist"
   ],
@@ -26,14 +26,15 @@
   "homepage": "https://turbo.hotwire.dev",
   "devDependencies": {
     "intern": "^4.9.0",
-    "rollup": "^2.9.1",
-    "rollup-plugin-typescript2": "^0.29.0",
+    "rollup": "^2.35.1",
+    "@rollup/plugin-node-resolve": "9.0.0",
+    "@rollup/plugin-typescript": "^6.0.0",
     "tslib": "^2.0.3",
     "typescript": "^3.8.3"
   },
   "scripts": {
     "clean": "rm -fr dist",
-    "build": "tsc -b && rollup -c",
+    "build": "rollup -c && tsc --noEmit false --declaration true --emitDeclarationOnly true --outDir dist/types",
     "watch": "rollup -wc",
     "start": "intern serveOnly",
     "test": "NODE_OPTIONS=--inspect intern"
