+++ src/tests/fixtures/additional_assets.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <title>Additional assets</title>
     <link rel="stylesheet" type="text/css" href="/fixtures/test.css">
-    <script src="/fixtures/turbo.js" data-turbo-track="reload"></script>
+    <script src="/fixtures/turbo.es2017-umd.js" data-turbo-track="reload"></script>
     <script src="/fixtures/test.js"></script>
   </head>
   <body>
