+++ src/tests/fixtures/async_script.html
@@ -8,7 +8,7 @@
       addEventListener("DOMContentLoaded", function() {
         setTimeout(function() {
           var script = document.createElement("script")
-          script.src = "/fixtures/turbo.js"
+          script.src = "/fixtures/turbo.es2017-umd.js"
           script.setAttribute("async", "")
           document.head.appendChild(script)
         }, 1)
