+++ src/tests/fixtures/async_script_2.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <title>Turbo</title>
     <script src="/fixtures/test.js"></script>
-    <script src="/fixtures/turbo.js" async></script>
+    <script src="/fixtures/turbo.es2017-umd.js" async></script>
   </head>
   <body>
     <section>
