+++ src/tests/fixtures/body_script.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <title>Body script</title>
-    <script src="/fixtures/turbo.js" data-turbo-track="reload"></script>
+    <script src="/fixtures/turbo.es2017-umd.js" data-turbo-track="reload"></script>
     <script src="/fixtures/test.js"></script>
     <meta name="turbo-cache-control" content="no-preview">
   </head>
