+++ src/tests/fixtures/one.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <title>One</title>
-    <script src="/fixtures/turbo.js" data-turbo-track="reload"></script>
+    <script src="/fixtures/turbo.es2017-umd.js" data-turbo-track="reload"></script>
     <script src="/fixtures/test.js"></script>
     <meta name="test" content="foo">
   </head>
