+++ src/tests/fixtures/tracked_asset_change.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <title>Tracked asset change</title>
-    <script src="/fixtures/turbo.js?123" data-turbo-track="reload"></script>
+    <script src="/fixtures/turbo.es2017-umd.js?123" data-turbo-track="reload"></script>
     <script src="/fixtures/test.js"></script>
   </head>
   <body>
