+++ src/tests/fixtures/turbo.es2017-umd.js
@@ -1,0 +1,1 @@
+../../../dist/turbo.es2017-umd.js
