--- src/tests/fixtures/turbo.js
@@ -1,1 +1,0 @@
-../../../dist/turbo.js
