+++ src/tests/fixtures/visit.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <title>Turbo</title>
-    <script src="/fixtures/turbo.js" data-turbo-track="reload"></script>
+    <script src="/fixtures/turbo.es2017-umd.js" data-turbo-track="reload"></script>
     <script src="/fixtures/test.js"></script>
   </head>
   <body>
