+++ src/tests/fixtures/visit_control_reload.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <title>Visit control: reload</title>
-    <script src="/fixtures/turbo.js" data-turbo-track="reload"></script>
+    <script src="/fixtures/turbo.es2017-umd.js" data-turbo-track="reload"></script>
     <script src="/fixtures/test.js"></script>
     <meta name="turbo-visit-control" content="reload">
   </head>
